pyproject.toml
[tool.poetry]
name = "verb-games-support-bot"
version = "0.1.0"
description = ""
authors = ["ilyashirko <ilyashirko@gmail.com>"]
readme = "README.md"
packages = [{include = "verb_games_support_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.14"
environs = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
